llm/generated_scans_evaluation/requirements.txt
torch==2.10.0
python-dotenv==1.2.1
regex==2026.1.15
accelerate>=1.1.0
bitsandbytes==0.46.1
deepeval==3.7.9
Pillow==10.4.0
tqdm==4.67.3
qwen-vl-utils==0.0.14
torchvision==0.25.0
openai==2.14.0
matplotlib==3.10.7
seaborn==0.13.2
scikit-learn==1.7.2